# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Celeste API
  description: API documentation for the Celeste e-commerce platform.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Inventory:
      type: object
      properties:
        id:
          type: string
          example: inv_123
        productId:
          type: string
          example: prod_123
        storeId:
          type: string
          example: store_123
        stock:
          type: number
          format: integer
          example: 100
        lastUpdated:
          type: string
          format: date-time
          example: '2024-07-25T10:00:00Z'
...
    UpdateInventory:
      type: object
      properties:
        stock:
          type: number
          format: integer
          example: 100
security:
  - bearerAuth: []
paths:
...
  /inventory/{id}:
    get:
      summary: Get an inventory item by ID
      tags:
        - Inventory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the inventory item to retrieve
      responses:
        '200':
          description: Inventory item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
        '404':
          description: Inventory item not found
    put:
      summary: Update an inventory item
      tags:
        - Inventory
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the inventory item to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInventory'
      responses:
        '200':
          description: Inventory item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
        '404':
          description: Inventory item not found
    delete:
      summary: Delete an inventory item
      tags:
        - Inventory
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the inventory item to delete
      responses:
        '200':
          description: Inventory item deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        '404':
          description: Inventory item not found
...